Create Peer Review Report
Authors should have the option to send out an Expression of Interest to academics who have registered themselves as Peer-Reviewers, inviting them to peer-review and comment on their Open Textbooks. Peer-Reviewers should later be able to download a report confirming they have undertaken Open Textbook peer-review activities, that they can add to their Google Profile or ORCID iD for the purpose of promotion or tenure. Is this possible?